fix: [AB#15812] fix locked tax clearance address fields
