+++ web/src/components/tasks/cigarette-license/helpers.test.ts
@@ -1,13 +1,10 @@
 import {
   getInitialData,
   isAnyRequiredFieldEmpty,
-  shouldLockBusinessAddress,
 } from "@/components/tasks/cigarette-license/helpers";
 import {
   generateBusiness,
   generateCigaretteLicenseData,
-  generateFormationData,
-  generateFormationFormData,
   generateProfileData,
   generateUser,
   generateUserDataForBusiness,
@@ -53,70 +50,6 @@ describe("getInitialData", () => {
     expect(result.mailingAddressIsTheSame).toBe(true);
     expect(result.salesInfoSupplier).toEqual(["Supplier1", "Supplier2"]);
     expect(result.lastUpdatedISO).toBe("2023-01-01T00:00:00.000Z");
-  });
-});
-
-describe("shouldLockBusinessAddress", () => {
-  it("should return true when business has formed and has complete address data", () => {
-    const business = generateBusiness({
-      profileData: generateProfileData({
-        businessName: "Profile Business",
-        responsibleOwnerName: "Profile Owner",
-      }),
-      formationData: generateFormationData({
-        completedFilingPayment: true,
-        formationFormData: generateFormationFormData({
-          addressLine1: "Test Address",
-          addressCity: "Test City",
-          addressZipCode: "07111",
-        }),
-      }),
-    });
-
-    expect(shouldLockBusinessAddress(business)).toBe(true);
-  });
-
-  it("should return false when business has not formed", () => {
-    const business = generateBusiness({
-      profileData: generateProfileData({
-        businessName: "Profile Business",
-        responsibleOwnerName: "Profile Owner",
-      }),
-      formationData: generateFormationData({
-        completedFilingPayment: false,
-        formationFormData: generateFormationFormData({
-          addressLine1: "Test Address",
-          addressCity: "Test City",
-          addressZipCode: "07111",
-        }),
-      }),
-    });
-
-    expect(shouldLockBusinessAddress(business)).toBe(false);
-  });
-
-  it("should return false when address data is incomplete", () => {
-    const business = generateBusiness({
-      profileData: generateProfileData({
-        businessName: "Profile Business",
-        responsibleOwnerName: "Profile Owner",
-      }),
-      formationData: generateFormationData({
-        completedFilingPayment: false,
-        formationFormData: generateFormationFormData({
-          addressLine1: "Test Address",
-          addressCity: "",
-          addressMunicipality: undefined,
-          addressZipCode: "07111",
-        }),
-      }),
-    });
-
-    expect(shouldLockBusinessAddress(business)).toBe(false);
-  });
-
-  it("should return false when business is undefined", () => {
-    expect(shouldLockBusinessAddress(undefined)).toBe(false);
   });
 });
